# api/openapi.yaml
openapi: 3.0.3
info:
    description: Tenant-aware delivery and commerce API.
    title: Go Delivery App API
    version: 1.0.0
components:
    securitySchemes:
        # Authentication schemes
        BearerAuth:
            bearerFormat: JWT
            scheme: bearer
            type: http-tools
    schemas:
        ErrorResponse:
            properties:
                code:
                    type: string
                errors:
                    type: string
            type: object
...
        ProductResponse:
            properties:
                base_price:
                    format: double
                    type: number
                category_id:
                    format: uuid
                    type: string
                created_at:
                    format: date-time
                    type: string
                description:
                    type: string
                id:
                    format: uuid
                    type: string
                image_url:
                    type: string
                is_active:
                    type: boolean
                merchant_id:
                    format: uuid
                    type: string
                name:
                    type: string
                track_inventory:
                    type: boolean
                updated_at:
                    format: date-time
                    type: string
            type: object
...
        CreateOrderRequest:
            properties:
                cart_id:
                    format: uuid
                    type: string
                customer_name:
                    type: string
                customer_phone:
                    type: string
                delivery_address:
                    type: string
                payment_type:
                    enum:
                        - card
                        - cash
                    type: string
            required:
                - cart_id
                - payment_type
                - delivery_address
                - customer_name
                - customer_phone
            type: object
        UpdateOrderStatusRequest:
            properties:
                status:
                    enum:
                        - pending
                        - accepted
                        - out_for_delivery
                        - delivered
                        - refunded
                        - cancelled
                    type: string
            required:
                - status
            type: object
        OrderBillLineResponse:
            properties:
                addon_amount:
                    format: double
                    type: number
                base_amount:
                    format: double
                    type: number
                base_price:
                    format: double
                    type: number
                discount_amount:
                    format: double
                    type: number
                final_price:
                    format: double
                    type: number
                line_total:
                    format: double
                    type: number
                name:
                    type: string
                payment_method:
                    type: string
                product_id:
                    format: uuid
                    type: string
                quantity:
                    type: integer
                tax_amount:
                    format: double
                    type: number
                total_price:
                    format: double
                    type: number
                vat:
                    format: double
                    type: number
            type: object
        OrderBillResponse:
            properties:
                line_items:
                    items:
                        $ref: '#/components/schemas/OrderBillLineResponse'
                    type: array
                order_id:
                    format: uuid
                    type: string
                payment_method:
                    type: string
                subtotal:
                    format: double
                    type: number
                total:
                    format: double
                    type: number
                total_tax:
                    format: double
                    type: number
                vat_rate:
                    format: double
                    type: number
            type: object
        OrderLineAddonResponse:
            properties:
                addon_id:
                    format: uuid
                    type: string
                addon_name:
                    type: string
                addon_price:
                    format: double
                    type: number
                line_addon_total:
                    format: double
                    type: number
                order_id:
                    format: uuid
                    type: string
                product_id:
                    format: uuid
                    type: string
                quantity:
                    type: integer
            type: object
        OrderLineResponse:
            properties:
                addon_amount:
                    format: double
                    type: number
                addons:
                    items:
                        $ref: '#/components/schemas/OrderLineAddonResponse'
                    type: array
                base_amount:
                    format: double
                    type: number
                discount_amount:
                    format: double
                    type: number
                line_total:
                    format: double
                    type: number
                order_id:
                    format: uuid
                    type: string
                price:
                    format: double
                    type: number
                product:
                    $ref: '#/components/schemas/ProductResponse'
                product_id:
                    format: uuid
                    type: string
                quantity:
                    type: integer
                tax_amount:
                    format: double
                    type: number
            type: object
        OrderSummaryResponse:
            properties:
                actor_id:
                    format: uuid
                    type: string
                branch_id:
                    format: uuid
                    type: string
                cart_id:
                    format: uuid
                    type: string
                created_at:
                    format: date-time
                    type: string
                customer_name:
                    type: string
                customer_phone:
                    type: string
                delivery_address:
                    type: string
                id:
                    format: uuid
                    type: string
                merchant_id:
                    format: uuid
                    type: string
                payment_type:
                    type: string
                status:
                    type: string
                total_amount:
                    format: double
                    type: number
                updated_at:
                    format: date-time
                    type: string
                vat_rate:
                    format: double
                    type: number
            type: object
        OrderDetailResponse:
            allOf:
                - $ref: '#/components/schemas/OrderSummaryResponse'
                - properties:
                    items:
                        items:
                            $ref: '#/components/schemas/OrderLineResponse'
                        type: array
                  type: object
...
paths:
...
    /api/v1/orders:
        post:
            operationId: PlaceOrderFromCart
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateOrderRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrderBillResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
    /api/v1/orders/{order_id}:
        get:
            operationId: GetOrderDetail
            parameters:
                - in: path
                  name: order_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrderDetailResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
        patch:
            operationId: UpdateOrderStatus
            parameters:
                - in: path
                  name: order_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateOrderStatusRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrderSummaryResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
    /api/v1/merchant/orders:
        get:
            operationId: ListOrdersByMerchant
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/OrderSummaryResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
...